move IEncryptionKeyProvider to separate file and rename EncryptionKeyProvider
